+++ src/flet_terminal/__init__.py
@@ -3,7 +3,7 @@
 from flet_terminal.extra_keys import DEFAULT_EXTRA_KEYS, ExtraKeysBar
 from flet_terminal.mobile_terminal import MobileTerminal
 from flet_terminal.search_bar import TerminalSearchBar
-from flet_terminal.terminal import Terminal
+from flet_terminal.terminal import ShortcutEvent, Terminal
 from flet_terminal.themes import BUILTIN_THEMES, get_theme
 
 __all__ = [
@@ -11,6 +11,7 @@ __all__ = [
     "DEFAULT_EXTRA_KEYS",
     "ExtraKeysBar",
     "MobileTerminal",
+    "ShortcutEvent",
     "Terminal",
     "TerminalSearchBar",
     "get_theme",